Convert speed to mile pace

speed_to_mile_pace converts a speed (in meters per second) to a mile pace. This method is vectorized, so it works on a column in a data frame. This is most useful after calling mutate_with_speed(), to convert that speed to the more-commonly-used pace. See vignette("pace") for examples.

Usage

speed_to_mile_pace(speed)

Arguments

speed

A vector of doubles representing speed in meters per second, as from mutate_with_speed().

Value

A corresponding vector of lubridate::duration values, representing the mile pace.

Examples

speed_to_mile_pace(3)
#> [1] "536.446666666667s (~8.94 minutes)"
speed_to_mile_pace(1)
#> [1] "1609.34s (~26.82 minutes)"